Handle menu encoder with the Attiny841
The menu encoder is currently handled by the mainboard in place of locally with
the Attiny841 MCU, which is not exactly ideal. The code here should be the first
part of the code migration to the Attiny841, the second part is to be done on
the mainboard which should process the new `GET_STATUS` command.