Encoder: Add software debounce for encoder switch
The hardware RC-filter that has been placed between the encoder
switch and the Attin841 pin works fine looking at the scope data,
but FSR the interrupt vector gets triggered multiple times (maybe
due to noise from the screen on the 5V line). It could be possible
that the schmitt-trigger is too sensative for noise and therefore
false trigger or is not available on the input interrupt (this is
not very clear in the datasheet).
This causes false button triggers, so adding an extra software
debounce filter should be enough to prevent these extra presses.