Update protocol code from latest bootloader
This copies the protocol code from a more recent bootloader, which has
been refactored to support I2C and RS485 (though no RS485 support for
the attiny exists, so code for RS485 is partially added but incomplete
and unused).

In addition to adapting to the new code structure, this adds support for
some of the bootloader commands (GET_PROTOCOL_VERSION, SET_ADDRESS and
GET_MAX_PACKET_LENGTH), which makes it easier to manage multiple child
boards from the mainboard.

This also renames Hardware.h to Config.h, which is what the bootloader
now uses for its configuration.

The following files were copied from ChildbusBootloader, commit 29f1e52
(PROTOCOL: Bump version to 2.1):

    cp ../../ChildbusBootloader/attiny/{TwoWire.cpp,Reset.cpp} src/attiny/
    cp ../../ChildbusBootloader/{BaseProtocol.cpp,BaseProtocol.h,Crc.h,Bus.h} src/